Add AgentThreadPool, responsible for producing advice
This allows to remove the use of self._advisors in the stable-baselines3
policies. Instead, a Policy gets a reference to a ThreadPool (every
thread for an agent_id), and the ThreadPool produces the advice vector.
This ensures that advice is always produced by the freshest list of
threads, without having to add every thread as an advisor to every
thread.