Use API keys instead of username/password to log into Shepherd
An API key is linked to a particular agent (and is supposed to be kept
secret by the user). Compromizing an API key can therefore be solved
easily by generating a new one for the agent, and destroying the
previous one. There is no need for a user-wide change of password.