Have a thread be advised only by threads that have learned recently
The idea is that a thread T, when it is created, loads the latest
snapshot of the agent as learned by any previous thread. Then, T starts
learning. It is interesting to advice it only with threads that are also
still learning, so that no outdated knowledge is used as advice.