Reuse threads that have finished episodes.
This makes Shepherd compatible with clients that execute only one or a
few episodes before stopping interacting with Shepherd. In these cases,
the RL agent is unable to learn anything from only a few time-steps, so
it is interesting that a thread (and its RL agent) can be re-used by a
future client, to continue acquiring experiences, eventually leading to
learning.